refactor: consolidate all Resend operations into src/resend/
Moves newsletter subscription helpers out of src/utilities/resend.ts
and broadcast send logic out of the inline handler into a dedicated
src/resend/ module, organized by operation domain.

- src/resend/client.ts    — shared Resend client, retry, rate-limit helpers
- src/resend/contacts.ts  — createResendContact, addContactToResendSegment
- src/resend/newsletter.ts — sendWelcomeEmail, handleNewsletterSubscribe
- src/resend/broadcasts.ts — createAndSendResendBroadcast (extracted from send.ts)
- src/resend/index.ts     — barrel re-export

handlers/send.ts now calls createAndSendResendBroadcast() instead of
instantiating new Resend() directly, picking up the shared client and
rate-limit retry logic. plugins/index.ts import path updated.
